+++ config.py
@@ -74,6 +74,6 @@ MAIL_USERNAME = os.getenv("VH7_MAIL_USERNAME", None)
 MAIL_PASSWORD = os.getenv("VH7_MAIL_PASSWORD", None)
 MAIL_DEFAULT_SENDER = os.getenv("VH7_MAIL_SENDER", None)
 
-# Recaptcha configuration
-RECAPTCHA_PUBLIC_KEY = os.getenv("VH7_RECAPTCHA_PUBLIC", None)
-RECAPTCHA_PRIVATE_KEY = os.getenv("VH7_RECAPTCHA_PRIVATE", None)
+# hCaptcha configuration
+HCAPTCHA_SITE_KEY = os.getenv("VH7_HCAPTCHA_SITE", "")
+HCAPTCHA_SECRET_KEY = os.getenv("VH7_HCAPTCHA_SECRET", "")